Deploy an example Node app to ECS
Add a Cloud Formation template, SAM config and a Hello World Node.js app
that can be deployed to run on AWS ECS using Fargate.

This is the smallest template that's needed to run an app and access it
via a DNS record from the public internet. Not all of this is needed for
every deployed service however - the networking stack (VPC, subnets,
internet gateway, routing etc.) is all something that can be set up once
per account and used across multiple services. The aim of this example
however is to work out everything that's needed to run an internet
facing service, so it's all in a single template and stack.

Once it's deployed with `sam deploy`, the service can be accessed at the
DNS name for the load balancer.

Note, the Node app is present in this repo but currently there's no
automation or instructions for building the container and pushing it to
ECR. Instead I built the container on my laptop and ran `docker push` to
push it to my account on Docker Hub. This isn't sustainable as Docker
Hub rate limits downloads, but it's fine for an example app. We have
followup work planned to set up a proper pipeline.