Add private subnets to shared VPC
Running EKS with Fargate as the compute engine requires the Kubernetes
control plane to run in a public subnet (so `kubectl` and similar
commands can reach the API), and the nodes on Fargate to run in private
subnets.

Change the common VPC to have two private and two public subnets. The
public subnets have access to the internet via an internet gateway, and
the private subnets route traffic through NAT gateways in the public
subnets. This allows resources in the private subnets access out to the
internet but prevents incoming connections.

Also update the fargate demo app template so it references the new
subnet outputs.