Set up an EKS cluster using Fargate compute
This commit adds the CloudFormation template and SAM config to deploy
an EKS cluster in the common VPC to run a 'Hello World' demo app.

The template also adds a managed IAM policy for the AWS Load Balancer
Controller. This is a Kubernetes service that creates a load balancer
pointing to a service on the cluster when a Kubernetes ingress is
applied. The additional tags on the subnets are used by this controller
to determine where to run the load balancers.

When this cluster is created, the only person who will have access is
the person who ran `sam deploy`. To grant others access, you'll need to
run `kubectl edit -n kube-system configmap/aws-auth`, add the role
arn of the role they assume in the AWS account and add that role to the
`system:masters` group.

You'll also need to follow the steps in [[1]] to set up a service
account and install the load balancer controller on the cluster.

[1]: https://docs.aws.amazon.com/eks/latest/userguide/aws-load-balancer-controller.html