Define a 'Hello World' app for EKS
Deploy the app to the cluster with `kubectl apply -f hello-world.yaml`.

Define a deployment of the same 'Hello World' Node app that's deployed
for the Fargate demo, a service to map TCP port 80 onto the container
and an ingress which uses the AWS load balancer controller to create a
load balancer which allows traffic to the service.