Add a DocumentDB cluster to the EKS stack
The credentials for this cluster come from AWS Secrets Manager and are
injected when the stack is created by CloudFormation using dynamic
references [[1]].

A secret with ID `hello-world-doc-db-credentials` and keys `admin-user`
and `admin-password` must be created before this change is deployed.

Create a subnet group for the cluster in our VPC's private subnets so
the Kubernetes pods (which run in the private subnets) can reach it.

[1]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/dynamic-references.html