Give hello-world pods read access to Secrets Manager
Create separate execution roles for the kube-system and hello-world pods
and give read access to Secrets Manager to the hello-worls pods.

Note that this policy gives read access to all secrets. This would not
be acceptable in a production system - we should limit access to only
the secrets the pod will need. In this proof-of-concept it's ok however.