Pass the database details to the app
as environment variables coming from a Kubernetes secret. This secret
has to be created before deploying the manifest, so add instructions to
the README for this.

Ideally the step of taking the CloudFormation outputs and setting them
as secrets would be done by a CD script, but we don't have that yet
so unfortunately it's a manual step for now