Add a basic Hello World app and use it in the manifest
Future changes to this app need the docker image to be built locally
and pushed to Dockerhub. Then the deployment needs to be restarted to
pull the new image:

`kubectl rollout restart deployment hello-world-deployment`

This will do a zero downtime rolling deploy and create new pods running
the latest image to replace any existing running ones.