Allow pods to assume the Execution Role
They're authenticated through the service account, which comes from the
cluster's OIDC provider.

Unfortunately the values in this condition have to come from the
cluster's OIDC provider which can only be done after the CloudFormation
is deployed and the cluster's been created. If we're creating a new
cluster we'll need to deploy the CloudFormation, set up the OIDC
provider and then update the values in this policy to the new ones.