Add a service to connect to DocumentDB
This commit adds the `mongodb` library as a dependency and creates a
new service that connects to the DocumentDB cluster using the
credentials fetched from Secrets Manager and the endpoint from the
environment variables.

Our DocumentDB cluster enforces TLS traffic, so this commit also adds
AWS's database public keys to the app.

See [[1]] for more explanation of the connection string parameters and
using TLS connections.

[1]: https://docs.aws.amazon.com/documentdb/latest/developerguide/connect_programmatically.html