Replace old MSK cluster with a new one
We hadn't allowed for plaintext communication with Kafka, which was
causing connection problems. We may want to tidy this up in future but
we really need to bootstrap this demo. So we've gone ahead with
plaintext.

CF was extremly unhappy about trying to make a breaking change to an MSK
instance outside of a maintianance window. So rather than try to
negotiate with it, we decided to create a new MSK cluster (with a stupid
integer on the end of the name) and prune the old one.

Seems to have worked!