Bring down Fargate demo app
I ran `gds aws di-solid-prototype -- sam delete` to bring the app down.
This commit just adds a note to the README to say it's down and gives
instructions on how to bring it back up if needed.

This is so we can keep the IaaC in this folder for reference when we're
doing future work.