Set up unit tests
Using Mocha, Chai and Sinon, and heavily inspired by the set up in
https://github.com/alphagov/di-authentication-account-management to
make it easier for developers to work on both repos.

As part of this, refactor the index router to have a controller file
(which can be easily unit tested) and a router which imports functions
from the controller. This is also the setup used in the account
management app.