Set up ECR registry for deployment
We're going to deploy this app to AWS ECS / Fargate, so we need to
build a Docker image and host that in a registry. We're already using
AWS so ECR is the sensible choice. We could use Dockerhub but we'll
probably hit the free account rate limiting problems at some point and
procuring a paid account will be more work than setting up ECR.

This commit also adds a bash script which replicates what our CD
pipeline will eventually do to build, tag and push the app image to
the new repository.