Deploy app to ECS
This isn't CD yet, but this commit adds a CloudFormation template and
SAM config to deploy the app to ECS / Fargate.

The CloudFormation is almost identical to what we used to deploy the
demo Fargate app, just with different resource names and image.

Note that this specifies to deploy only one container. Initially we'll
be using in-memory session management, so having multiple deployed
containers brings the risk of requests being split between containers
which will break as only one will have the session state.

I *think* it's possible to get the load balancer to do session-aware
routing, but until we've confirmed that or set up an external session
store we should keep this to a single container.