Update deploy instructions
Rename `build_and_push_image.sh` to `deploy.sh` and add instructions
to the README for how to use it.

Also remove the default image tag from the CloudFormation template.
This is passed in by `deploy.sh` and I don't want to leave in a very
old default image and accidentally deploy it.