GUA-230: Set up client side sessions with cookie-session
In a production app we'd want to use server side session management with
a backing service like Redis, but storing the sessions on the client
side is fine for this prototype and brings some advantages over doing
in-memory server side sessions.

As long as the signing keys stay the same, with client side sessions we
can deploy multiple instances of the app behind a load balancer and
they'll all be able to decode a session. We also won't lose sessions
when we deploy a new version of the app and restart all the containers.

Once we start adding JWTs to the session state we'll need to keep an
eye on the size of the cookie - we can't go above 4Kb due to browser
limits. If that happens we'll need to change our session strategy, but
this is fine for now.

Author:    Alex Whitehead-Smith <alex.whitehead-smith@digital.cabinet-office.gov.uk>