GUA-230: Add login via `start.inrupt.com`
Nothing happens after the login, but this sets up the Inrupt Node client
library and adds some routes to login and handle the callback to
exchange the access code for an ID token.

This is mostly following [Inrupt's tutorial](https://docs.inrupt.com/developer-tools/javascript/client-libraries/tutorial/authenticate-nodejs-web-server/)

The notable change is in the callback handler - we need to use
`req.originalUrl` when we pass the state to handle the incoming redirect.
This is because these routes are in a separate router that mounts
`/login`, so `req.url` is only `/callback` here and not `/login/callback`
which is what the app is expecting.

Author:    Alex Whitehead-Smith <alex.whitehead-smith@digital.cabinet-office.gov.uk>