GUA-230: Read and write from a pod
Sets up a working example where a user can login, click
a button to write an event to an access log stored in their pod
then view that access log (in the server console)

There's also some middleware at the access log router which forces
someone to be logged in.

A user can go through the journey (start at `/login`) and
click to write an event to their pod. Then when they click the button
again they can see the last access event in the console log and it'll
overwrite that with a new event

Co-authored-by: pmcb55 <patm@inrupt.com>