Try using aws-actions/aws-cloudformation-github-deploy
Previous attempts seemed to hit a brick wall. We wanted a task
definition to follow along with that github's guides wanted. They
suggsted that should be introduced as-code defined somewhere like
.aws/task-definition.json.

However we've defined ours in cloud formation, and I can't see a way to
supply it with the version that exists in AWS, and get it to pull JSON.

Looking at Alex's working deploy.sh though, we can perhaps skip all
that.

All we want to do is:
1. build a docker image (working!)
2. push it to ECR (working!)
3. sam deploy

I think this is what the new element in this commit should help with