GUA-220: Give CI pipeline full AWS access
We went back and forth on this, but after we'd spent about 30 minutes
playing permissions whack-a-mole with no clear end in sight we decided
to give the CI pipeline permission to modify any AWS resources.

This is where we want to end up anyway, with us doing very few manual
deploys, and we have all the appropriate security controls in place for
what actions run and when.