GUA-205: Render a client ID document
This will be needed so we can add it to the allow list in ESS for apps
which can create / modify pods, WebID profiles etc.

This puts the client ID document on `/info/id`. I thought about
following Inrupt's example of `/app/id` but then I had an `appRouter`
which was very similarly named to the entry point to the app - `app.ts`.
I think that has the potential to be confusing, so I went with `info`
instead.

Also adds some tests to make sure the client ID document has the
minimum required values.