GUA-205: Include local callback URI
This is to allow local development.

The client ID document needs to be on a URI that's accessible to our
ESS instance so it can fetch the document and validate it.

This won't work while we're doing local development, so we'll need to
add the localhost callback URL to the list of allowed redirect URIs
and provide the deployed app's client ID document when the local app
logs in. This means ESS will check the deployed app's document and
see that it's allowed to redirect users back to localhost while we're
developing the app.