GUA-230: Allow fetching a refresh token
Update the Solid-OIDC client ID document to include the scopes and
grant types that allow the app to request a refresh token.

This is because we need a refresh token in order to make further calls
to ESS on pages after the login callback. The Solid client library will
handle including the refresh token in any further calls of `Session.fetch`
so this should be all that's needed to get the read / write functionality
in #37 working again.