GUA-230: Display last accessed time on a page
When a user first comes to the `/access-logs` page, check if they
already have a log event in their pod. If they do, display the
timestamp.

If not, handle the `fetchError` and display an empty state message.

This means we can clean up the POST handler as it now only needs to
create a dataset if it doesn't exist and write the event to it.

Note: This 'access log' is slightly different to what we'll eventually
want to implement for the log we display in the account area. This is a
single `Thing` in the dataset which has a single timestamp property
which we overwrite each time.

When we build the actual access log we'll need to have a dataset with
one `Thing` per event that will each have a `DCTERMS.created` timestamp
so we can maintain the history.