GUA-210: static views for account home & subpages
Build static frontend templates and views of the account dashboard and
account subpages as illustrated in the Figma prototype.

These views are the basis for the user's dashboard where they can
view their data. Currently the data is hard coded, but the intention is
that the dashboard will interact with the user's solid pod and display
their identity attributes, communication preferences and existing access
grants.