Make getOrCreateDataset & getDatasetUri reusable
We'll shortly want to re-use these methods for reading / writing a VC.
Also make the path we're writing to a pod something we pass it.

Along the way I've also sneakily changed where we're storing the Access
logs. I have an idea that a consistent pathing alongside the VCs might
make it easier to reason about and something like:

`private/` - not public data
`govuk/` - GOV.UK stuff, not anywhere else in gov
`identity/` - Our program / unit
`poc/` - What the service is

and then the resource name.. makes sense