GUA-209: Remove unneeded interfaces
We don't need interfaces to manage the structure of the JWT - the
`jsonwebtoken` library constructs the object and converts it to a
string in one go.

Therefore, we can delete these interfaces and make the VC constructors
return a `Credential` instead - this interface was identical to the
`IdentityCheckCredential` anyway.

Also, rename the interface to `Credential` since it's a single check
not many.