GUA-247: Check user's pod for previous stored credentials
When a user first starts the IPV journey, check their pod for any
previously stored credentials. If both the credentials are there, take
them past the questions to the re-use page.

Note: We're not doing any validation of these credentials, or even
checking the binary JWTs are present - we're only looking to see that
there's resources at the metadata paths where we're expecting them to
be. This is a deliberate choice made to speed up the development of
this PoC; in the real thing we'd need to fetch the JWTs, decode them
and run validity checks or re-use the data for other CRIs.