GUA-231: Move login login into GET route
This removes the intermediate login screen users saw previously. This
screen was helpful when we were debugging logins, but now it's working
there's no need for it and it's not in our designs.

Now we can seamlessly send users to login by redirecting them to this
route, optionally passing a `returnUri` URL parameter with a URI to
send them back to after they've logged in. For example in a controller
or router middleware:

```js
res.redirect(`$getHostname()/login?returnUri=/identity`
```