GUA-231: Send users back to the page they're on after login
When a user tries to visit a page that requires login and isn't logged
in, the middleware intervenes and redirects them to login.

Now we're using more than one journey, we need to send users back to
different pages after they login. Pass the `req.originalUrl` through
the middleware to `/login` to send them back to the originally requested
page.