BAU: Close login request once user is sent to IDp
This was causing the app to crash when a user spent more than 3.5
seconds logging into their account as that's the default request
timeout length in Express.

Changing the login function to not `await` for the response from
`session.login` closes the request once the user's been redirected to
the IDp. This has no impact on how the login works or what the user
sees.

This is the only part of the `loginGet` function that's async so we can
also make the whole function synchronous now.