GUA-231: Retrieve existence of credentials in Home
Use it to vary the page either offering the user a chance to go prove
and save their ID, or give them the "View" and Delete button.

NB contents of table are static. In a later PR we can consider if it's
worth saving RDF that:
- Puts a created_at Timestamp
- Puts a locale specific name

And then read both out into the table.