GUA-231: Connect ID reuse journey branch to service journey
This commit adds the 'Your identity has been confirmed' page which
comes after the user is asked to re-use the stored identity, and fills
in the service link and name to the template from the user's session.

It sets up the routing to pass a user back to the correct mock service
journey after they've completed the IPV journey. When the DBS journey
passes a user to login and the IPV journey, it writes the service
name and return URL into the user's session.

The service name is used to populate templates at the end of the IPV
journey, eg. sections like 'continue to apply for a basic DBS check'.

The return URL is the URL of the next page in the DBS check journey.

When we add the tax account journey all we need to do is populate a
similar object in the session when that mock service passes a user to
IPV, and this logic will send the user onto the correct place.