GUA-314: Save NI number to user's pod
Construct a VC in the same way as we are for the various ones in the
IPV journey and save it to a user's pod once they give consent.

I've taken a bit of liberty here in defining the NI number as a
property that's defined in a DWP-owned vocab, but it demonstrates the
spirit of cross-government collaboration.