BAU: make dbs/pta start pages look more like GOVUK
Make our fake "GOVUK" pages in the app look a bit more like GOVUK and a
bit less like a service page through removal of the phase banner,
addition of breadcrumbs and minor spacing tweaks.

Not adding the super nav to these pages for now as it would add a lot of
CSS and JS to the payload for very little benefit.