GUA-311: Create a function to initate DWP access requests
This function follows the flow detailed in Inrupt's Manage Access
Request docs:

https://docs.inrupt.com/developer-tools/javascript/client-libraries/tutorial/manage-access-requests/#requesting-access

The pattern is this, DWP are going to construct an **Access Request**,
this will then be sent for to the user (almost like a contract), if the
user consents the terms set out here will be agreed to.

NB: This isn't all or nothing, under some circumstances Solid can allow
for a partial acceptance, see here for more detail:
https://docs.inrupt.com/developer-tools/api/javascript/solid-client-access-grants/modules/manage.html#approveaccessrequest

But for now, lets keep it simple, we're going all in.

DWP gets to set a terms of expiry of access, in this case we just want
to write once, so lets give ourselves a generous five mins to complete
everything (We'll probably have to consider what happens if the user
times out / what a reasonable session length is).

We then need to set out our terms, including:
- What kind of access we want (read, write, append)[1][2]
- The WebID of the person who owns the resource (the user)
- Which resources we're talking about [3]
- Purposes[4]

Then we ping the request to the VC endpoint (in this case the one in ESS
will do as that's a common link between our user and the DWP app) and
make the request using the **DWP session**.

[1]: https://github.com/inrupt/solid-client-access-grants-js/blob/0fe35cdaa04d0a14e4167b664194b00a9a767b33/src/type/AccessModes.ts#L33-L35
[2]: https://docs.inrupt.com/ess/latest/security/acp/?highlight=append#access-modes
[3]: NB: There's a gotcha here, note that what I'm passing in is the
container we want to put the national insurance number into. Initially I
was imagining a resource! But one doesn't exist yet, so we were getting
404s. In this case I'm granting the app permission to write to the
container, not to write to a resource that doesn't exist yet!
[4]: These are super interesting from a GDPR perspective, right now I'm
just going to pass these as a URL, but I think these are an excellent
candidate for hosting within the app as Linked Data. We can imagine a
reimplientation of Pat's Multilingual PR:
https://github.com/alphagov/di-solid-prototype/pull/76 where DWP, or any
department have hosted all the reasons they thing they have to
manipulate data, hosted on a URL, with multilingual content. This would
allow the access management app to then pull in Human readable (one
hopes!) processing purposes, allowing us to be really clear with users
about why this data transfer is being undertaken.