GUA-311: Add helper functions, account can decode Access grants
Back in the "Account", we're going to imagine a new account feature
which handles access management. It is going to recieve Base64 encoded
"Access Requests", which are JSON-LD structured Verifiable Credentials.

It's job is going to be to:
1. Validate the request, ensure it is properly formatted and from
   someone who exists
2. Put the terms of the request to the user, see if they accept or deny
   the request
3. Respond to the request, so the VC store can be updated.

Here we start that by adding a small helper method to decode the Base64
encoded VCs, and a few checks to appease typescript along the way