GUA-311: Create requested container if not present
The approve / deny access request functions seem to check if the
resources mentioned in the access request actually exist, and fail with
a `FetchError` if not.

This is a bit of a hack which creates an empty container at the
requested URI before handling the access request to get the PoC
working.