GUA-311: Only create resource if missing
Previously we would try to write back a resource in any case, even if
it already existed. In the case where the resource existed but was
empty, trying to write back an unchanged empty dataset gave a 409
Conflict response.

Change the logic so we only try to write a new empty dataset if there
isn't one there already