GUA-311: Request permissions to resources directly
Something strange is going on when the DWP app is trying to write a
new dataset into an empty container, but I can't work out what.

Requesting each resource directly means the access management app
creates empty datasets at these locations first, which the DWP app can
then update. This gets it working consistently for the PoC which is
good enough for now.