GUA-324: Add logout buttons to the homepage
This is a bit of a hack because users of the PoC app are logged in to
two separate external services - DI's Auth and the ESS broker.

Both support OIDC standards compliant logout, but it'll be a bit
complicated to set up and we'd need to get someone on Auth or DfA to
add a logout redirect URL to the client registry for us.

Instead, simply add two buttons on the homepage which each log the
user out of one external service.