GUA-316: Add an interface to make access grant data readable
Currently we're displaying the URIs provided in the access grant on
the page in the account which allows people to approve or deny them.
This isn't ideal as they're hard to read and understand.

They are, however, all URIs so in the real world they'd resolve to
linked data documents which would include a name property which we
could fetch and display to the user.

In this prototype that's not the case, so we need a workaround.

This commit adds a class which acts as an interface to abstract the
details of fetching the names away from the controller. In this case,
I've just hard coded the values but this design allows us to swap out
the workings to a different solution, such as fetching from documents
or reading from the app's locales files instead.