GUA-250: Add journey flowcharts
These flow charts are simplified versions of the screens a user would
see, with enough of the 'what's going on behind the scenes' added to
make them useful in explaining the prototype.

Between the screens in Figma, these diagrams and the sequence diagrams
that Max added, we should have information for any level of technical
interest.